Remove useless casts to int
In an expression of the form:
[unsigned int] = (int)[size_t];
the cast to int serves only to invoke implementation-defined behavior,
since size_t is unsigned.  Remove such casts from mdc2dgst.c.

Also change the num element of struct mdc2_ctx_st to be a size_t instead
of unsigned int, so as to have the assignments mentioned above store to
a variable of the same width.

+++ crypto/mdc2/mdc2dgst.c
@@ -94,7 +94,7 @@ int MDC2_Update(MDC2_CTX *c, const unsigned char *in, size_t len)
         if (i + len < MDC2_BLOCK) {
             /* partial block */
             memcpy(&(c->data[i]), in, len);
-            c->num += (int)len;
+            c->num += len;
             return 1;
         } else {
             /* filled one */
@@ -112,7 +112,7 @@ int MDC2_Update(MDC2_CTX *c, const unsigned char *in, size_t len)
     j = len - i;
     if (j > 0) {
         memcpy(&(c->data[0]), &(in[i]), j);
-        c->num = (int)j;
+        c->num = j;
     }
     return 1;
 }
